Add oi's own little optional language.
An altenative syntax for C when writing oi code.

The amount of code in this new language is marginally smaller; since oi
already has a low boiler-plate level.

The compiler is a pre-processor that generates C code, The preprocessor can
also act as a compiler; that pre-processes all .c files provided on a
commandline before calling gcc.